pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dspy-workspace"
version = "0.1.0"
urls = {"Homepage" = "https://github.com/AcidicSoil?tab=repositories", "Repository" = "https://github.com/AcidicSoil/dspy_workspace"}
description = "Collection of DSPy-based tools, configs, and tutorials"
readme = "README.md"
requires-python = ">=3.9,<3.13"
license = {text = "TBD"}
authors = [{name = "TBD"}]
dependencies = [
  "dspy",
  "requests",
  "python-dotenv",
  "beautifulsoup4",
  "html2text",
  "pydantic",
  "llm-ctx @ git+https://github.com/AnswerDotAI/llm-ctx.git",
  "llms_txt",
]

[project.optional-dependencies]
web = [
  "fastapi",
  "uvicorn",
]
dev = [
  "mlflow",
  "openai",
  "litellm[proxy]",
  "vllm",
  "datasets<4.0.0",
  "ollama",
  "orjson",
  "pre-commit",
  "ruff",
  "pytest",
]

[tool.hatch.build]
include = [
  "*.py",
  "*/**/*.py",
  "*/**/*.html",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
addopts = "-ra"

[tool.hatch.metadata]
allow-direct-references = true
